hematologia-clinic/backend/requirements.txt
# ================================================================
# Backend Dependencies — Hematología Clinic
# ================================================================

# ─── Framework ───
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# ─── Base de datos (async) ───
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ─── Configuración ───
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# ─── Seguridad / Auth ───
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0          # TOTP para 2FA

# ─── Redis ───
redis[hiredis]==5.1.1
celery[redis]==5.4.0
flower==2.0.1

# ─── Elasticsearch ───
elasticsearch[async]==8.15.1

# ─── MinIO ───
minio==7.2.10

# ─── Email ───
aiosmtplib==3.0.2
jinja2==3.1.4         # Templates de email

# ─── PDF ───
reportlab==4.2.5      # Generación de PDFs de informes
Pillow==11.0.0        # Procesamiento de imágenes (firma, logo)

# ─── WhatsApp ───
httpx==0.27.2         # Cliente HTTP async (para WhatsApp API)

# ─── Logging ───
structlog==24.4.0

# ─── Tests ───
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2         # Test client
factory-boy==3.3.1    # Factories para tests

# ─── Linting ───
ruff==0.7.1
mypy==1.13.0
